requirements.txt
annotated-types==0.7.0
anyio==4.6.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
msal==1.31.0
numpy==2.1.1
Office365-REST-Python-Client==2.5.12
openai==1.47.1
openpyxl==3.1.5
pandas==2.2.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywin32==306
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
